Check for int/uint value before encoding/after decoding that it actually fits into the desired type

+++ test/index.js
@@ -174,7 +174,7 @@ describe('decoding int32', function() {
     assert.equal(a.length, 1)
     assert.equal(a[0].toString(), b.toString());
 
-    var a = abi.rawDecode('neg', [ 'int32' ], [ 'int32' ], new Buffer('ae4f88b1ffffffffffffffffffffffffffffffffffffffffffffffffffffb29c26f344fe', 'hex'))
+    var a = abi.rawDecode('neg', [ 'int32' ], [ 'int64' ], new Buffer('ae4f88b1ffffffffffffffffffffffffffffffffffffffffffffffffffffb29c26f344fe', 'hex'))
     var b = new BN(-85091238591234)
     assert.equal(a.length, 1)
     assert.equal(a[0].toString(), b.toString())
